+++ lib/src/inputs/src/pickers/date/single.dart
@@ -75,6 +75,12 @@ class ThemedDatePicker extends StatefulWidget {
   /// By default, it will use `BorderRadius.circular(10)`.
   final BorderRadius borderRadius;
 
+  /// [errors] is the list of errors of the input.
+  final List<String> errors;
+
+  /// [hideDetails] is the state of hiding the details of the input.
+  final bool hideDetails;
+
   /// [ThemedDatePicker] is a date picker input. It is a wrapper of [ThemedTextInput] with a date picker.
   const ThemedDatePicker({
     super.key,
@@ -104,6 +110,8 @@ class ThemedDatePicker extends StatefulWidget {
     this.splashColor = Colors.transparent,
     this.highlightColor = Colors.transparent,
     this.borderRadius = const BorderRadius.all(Radius.circular(10)),
+    this.errors = const [],
+    this.hideDetails = false,
   }) : assert((label == null && labelText != null) || (label != null && labelText == null));
 
   @override
@@ -149,6 +157,8 @@ class _ThemedDatePickerState extends State<ThemedDatePicker> {
       disabled: widget.disabled,
       readonly: true,
       onTap: widget.disabled ? null : _showPicker,
+      errors: widget.errors,
+      hideDetails: widget.hideDetails,
     );
   }